Import PortAudio audio library v190600_20161030.
Builds on Linux via CMake. Not yet tested on building on Windows.
This release is the latest, which is from Octover 30, 2016.

As this project seems to be slowly updated, I have opted to import
the code from it directly, instead of having a Git submodule.
I don't intend that this code will need to be constantly updated.

PortAudio is MIT licensed, as found in its LICENSE.txt file.