Building synthesizer with PortAudio library. Only outputs sine wave.
CMakeLists.txt now looks for the portaudio library under /usr/local.
This should make it able to be found on OSX with Homebrew or MacPorts.

PI constant changed to M_PI, which is found in <cmath>.
An #ifndef define in sound.h if it isn't found in <cmath>.

New Sound class that currently only outputs a Sine wave in mono.
New ScopedPaHandler class to terminate PortAudio when out of scope.

olcNoiseMaker.h is no longer used.
Code that used it is temporarily commented out via #if 0 / #endif.

The synth code is not currently hooked up to PortAudio.
The MIDI code is also temporarily commented out.

May possibly build on Windows, but only Unix has been tested so far.