Generate basic Juce-based Audio Application via the Projucer.
All code in the directory JuceLibraryCode is from the Juce Project
Version 5.4.5. It is not a part of MKUltra.
Its presence in this repository is to make building MKUltra easier.
The Juce modules juce_audio_basics, juce_audio_devices,
juce_blocks_basics, juce_core and juce_events are ISC licensed.
The remaining Juce modules are GPLv3/Commercial licensed.

The project code for MKUltra does not yet have a license.

Currently there are 5 build targets for MKUltra:
- Android (via Gradle)
- Linux   (via Makefile)
- MacOSX  (via XCode)
- iOS     (via XCode)
- Windows (via Visual Studio 2019)

These may be removed at a later date. The Windows target is primary.

Open up the MKUltra.sln file in the Builds/VisualStudio2019 folder with
Visual Studio 2019 then click on Start Debugging to run the code.