Implement simple sine waves, triggered via MIDI inputs.
Heavily based on the Juce tutorial "Build a MIDI Synthesizer".

Basic GUI of a MIDI keyboard, which also captures keystrokes from
a PC keyboard.

Sine wave is real-time calculated, but sounds good.
The sine wave code is encapsulated in the SineWave.h file.
ADSR is crudely implemented, without any external controls.
This will be fleshed out using the previous Envelope code.

The classes in SineWave.h can be used as a reference basis for
other waveforms, such as triangle and square waves.

Currently there is 4 voice polyphony, but it can be increased.

The Synth.h file encapsulates the SynthAudioSource class, which
is the source of the audio output from the synthesizer.
Voices must be registered with the SynthAudioSource to be used.