Add SineWave.h and Synth.h to Projucer project.
From now on, the Projucer will be used to maintain the project,
not the IDEs such as Visual Studio.
The file MKUltra.jucer is the Projucer configuration file, and
everything in the Builds directory will be overwritten when the
MKUltra.jucer file is saved with the Projucer.

For example, adding a new file to the project must be done via
the Projucer, or else it will be overwritten.